project/requirements.txt
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
apache-airflow-providers-amazon

pandas
numpy
kafka-python     
watchdog 
plotly==6.3.1 
matplotlib==3.9.4
streamlit==1.50.0 
psycopg2-binary